bus: move subscriber
