+++ internal/test/e2e/cluster.go
@@ -17,13 +17,13 @@ import (
 	"go.sia.tech/core/consensus"
 	"go.sia.tech/core/types"
 	"go.sia.tech/coreutils"
+	"go.sia.tech/coreutils/chain"
 	"go.sia.tech/jape"
 	"go.sia.tech/renterd/api"
 	"go.sia.tech/renterd/autopilot"
 	"go.sia.tech/renterd/build"
 	"go.sia.tech/renterd/bus"
 	"go.sia.tech/renterd/config"
-	"go.sia.tech/renterd/internal/chain"
 	"go.sia.tech/renterd/internal/node"
 	"go.sia.tech/renterd/internal/test"
 	"go.sia.tech/renterd/internal/utils"
@@ -67,7 +67,6 @@ type TestCluster struct {
 
 	network *consensus.Network
 	cm      *chain.Manager
-	cs      *chain.ChainSubscriber
 	apID    string
 	dbName  string
 	dir     string
@@ -310,7 +309,7 @@ func newTestCluster(t *testing.T, opts testClusterOptions) *TestCluster {
 	tt.OK(err)
 
 	// Create bus.
-	b, bSetupFn, bShutdownFn, cm, cs, err := node.NewBus(busCfg, busDir, wk, logger)
+	b, bSetupFn, bShutdownFn, cm, err := node.NewBus(busCfg, busDir, wk, logger)
 	tt.OK(err)
 
 	busAuth := jape.BasicAuth(busPassword)
@@ -368,7 +367,6 @@ func newTestCluster(t *testing.T, opts testClusterOptions) *TestCluster {
 		logger:  logger,
 		network: busCfg.Network,
 		cm:      cm,
-		cs:      cs,
 		tt:      tt,
 		wk:      wk,
