+++ internal/utils/chain.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"time"
+
+	"go.sia.tech/core/types"
+)
+
+func IsSynced(b types.Block) bool {
+	return time.Since(b.Timestamp) <= 3*time.Hour
+}
