+++ stores/chain.go
@@ -4,13 +4,9 @@ import (
 	"context"
 
 	"go.sia.tech/core/types"
+	"go.sia.tech/coreutils/chain"
 	"go.sia.tech/coreutils/wallet"
-	"go.sia.tech/renterd/internal/chain"
 	"go.sia.tech/renterd/stores/sql"
-)
-
-var (
-	_ chain.ChainStore = (*SQLStore)(nil)
 )
 
 // ChainIndex returns the last stored chain index.
@@ -24,7 +20,7 @@ func (s *SQLStore) ChainIndex(ctx context.Context) (ci types.ChainIndex, err err
 
 // ProcessChainUpdate returns a callback function that process a chain update
 // inside a transaction.
-func (s *SQLStore) ProcessChainUpdate(ctx context.Context, applyFn chain.ApplyChainUpdateFn) error {
+func (s *SQLStore) ProcessChainUpdate(ctx context.Context, applyFn func(sql.ChainUpdateTx) error) error {
 	return s.db.Transaction(ctx, func(tx sql.DatabaseTx) error {
 		return tx.ProcessChainUpdate(ctx, applyFn)
 	})
@@ -32,7 +28,7 @@ func (s *SQLStore) ProcessChainUpdate(ctx context.Context, applyFn chain.ApplyCh
 
 // UpdateChainState process the given revert and apply updates.
 func (s *SQLStore) UpdateChainState(reverted []chain.RevertUpdate, applied []chain.ApplyUpdate) error {
-	return s.ProcessChainUpdate(context.Background(), func(tx chain.ChainUpdateTx) error {
+	return s.ProcessChainUpdate(context.Background(), func(tx sql.ChainUpdateTx) error {
 		return wallet.UpdateChainState(tx, s.walletAddress, applied, reverted)
 	})
 }
